+++ lib/src/ui/global/custom_base_view_model.dart
@@ -1,3 +1,5 @@
 import 'package:stacked/stacked.dart';
 
+/// [CustomBaseViewModel] is where shared variables and method that are used by
+/// the `ViewModels` live.
 abstract class CustomBaseViewModel extends BaseViewModel {}