+++ lib/src/ui/views/contributors/contributors_view.dart
@@ -1,19 +1,27 @@
-import 'dart:async';
-
-import 'package:contributors/src/app/utils/utils.dart';
+import 'package:flutter/material.dart' show ListTile;
 import 'package:flutter/material.dart';
 import 'package:github/github.dart';
 import 'package:stacked/stacked.dart';
 
+import 'package:contributors/src/app/utils/utils.dart';
 import 'package:contributors/src/ui/widgets/contributor_list_tile.dart';
 
 import 'contributors_view_model.dart';
 
+/// [ContributorsView] is the ready to use UI that lists all contributors of
+/// provided repository, ([ownerName]/[repoName]), showing there Name, Avatar,
+/// and Contributions Number in a [ListTile].
 class ContributorsView extends StatelessWidget {
+  /// [ownerName] is the name of the owner of [repoName].
   final String ownerName;
+
+  /// [repoName] is repo name that the contributor contributed to.
   final String repoName;
+
+  /// [locale] is the [Locale] that you want the text to appear in.
   final Locale locale;
 
+  /// `ContributorsView` is the default contructor for [ContributorsView].
   const ContributorsView({
     Key key,
     @required this.ownerName,
@@ -50,7 +58,7 @@ class ContributorsView extends StatelessWidget {
                 final List<Contributor> contributorsList = snapshot.data;
 
                 if (contributorsList == null || contributorsList.isEmpty) {
-                  return Center(
+                  return const Center(
                     child: CircularProgressIndicator(),
                   );
                 }