+++ lib/src/ui/widgets/contributor_list_tile.dart
@@ -1,13 +1,26 @@
 import 'package:contributors/src/app/utils/utils.dart';
 import 'package:flutter/material.dart';
 
+/// [ContributorListTile] is a widget used to show one contributor info when
+/// using a [ListView] to show basic info about him and how many contributions
+/// he/she did.
 class ContributorListTile extends StatelessWidget {
+  /// [imageUrl] is the image url for the contributor.
   final String imageUrl;
+
+  /// [name] is the name of the contributor.
   final String name;
+
+  /// [contributionsNumber] is the number of contributions of the contributor.
   final int contributionsNumber;
+
+  /// [onTap] is the image url for the contributor.
   final void Function() onTap;
+
+  /// [locale] is the [Locale] that you want the text to appear in.
   final Locale locale;
 
+  /// `ContributorListTile` is the default contructor for [ContributorListTile]
   const ContributorListTile({
     Key key,
     @required this.imageUrl,
@@ -25,7 +38,7 @@ class ContributorListTile extends StatelessWidget {
       ),
       title: Text(name),
       subtitle: Text(
-        "${Utils.getCurrentTranslation(locale).contributions}: $contributionsNumber",
+        '${Utils.getCurrentTranslation(locale).contributions}: $contributionsNumber',
       ),
       onTap: onTap,
     );