+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - Basic functionality of getting contributors of a repository
 - Support localization
+- Add `lint` package for linting